backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic>=2.11.2
pydantic-settings==2.6.0

# Session signing (PyJWT used for session cookies)
PyJWT[crypto]==2.10.1

# HTTP
requests==2.32.3
httpx>=0.27.0

# Environment variables
python-dotenv==1.0.1

# Data processing (numpy used by embedding_utils)
numpy>=1.24.0

# AWS SDK (for DynamoDB)
boto3>=1.35.0
# Rate limiting
slowapi>=0.1.9

# Language detection (for news deduplication scoring)
langdetect==1.0.9

# PDF parsing (for Civic Scout council document extraction)
llama-parse>=0.6.0

# Supabase / PostgreSQL (asyncpg connection pool for Supabase adapters)
asyncpg>=0.30.0
# Supabase Python client (async admin API for auth.users email lookup)
supabase>=2.0.0

# Stripe (license key management)
stripe>=8.0.0
